+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # Changelog
+## 1.0.1 (04/05/2020)
+
+## Improvements
+ * Added 1_NftStaking.js migration script
+
+
+## 1.0.0 (03/05/2020)
+
+### Breaking changes
+ * Migration to `@animoca/ethereum-contracts-core_library:1.0.0` with `solc:0.6.x` and `@openzeppelin/contracts:3.x`.
+
+### New features
+ * Added `NftStakingMock.sol` and `NftStakingTestableMock.sol`.
+
+### Improvements
+ * Better abstraction of core staking features.
 
 ## 1.0.0 (03/05/2020)